Extract the TV discover feed analytics into a reusable tracker
